// src/models/learners/learner.ts
export default interface LearnerModel {
  userId: string;
  username: string;
  about?: string;
  country?: string;
  displayname?: string;
  email?: string;
  userCategory: string;
  firstname: string;
  lastname?: string;
  gender?: string;
  referenceId?: string;
  rosterGlobalUserId?: string;
  schoolDistrict?: string;
  thumbnail?: string;
}